Add get_group_by_cols for Tuple.
Irrespective of whether we have an aggregate, we want to drill down to
the children here. You can't GROUP BY a tuple like you would for a
"normal" function - i.e. GROUP BY ("a", "b") is invalid SQL. However,
it's logically equivalent to GROUP BY "a", "b" in call cases, so we
never need the case where this 'function' needs to be included in the
clause.